feat: added responsive design
