--- src/GlobalStyling/GlobalStyle.css
@@ -1,31 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Carrois+Gothic&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Inconsolata:wght@200;300;400;500;700&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,600;1,200;1,400;1,600&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Junge&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Dancing+Script:wght@400;600&display=swap');
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-a {
-  text-decoration: none;
-  color: black;
-}
-h1 {
-  font-family: 'Junge', serif;
-}
-h2,
-h3,
-h4 {
-  font-family: 'Inconsolata', monospace;
-}
-body {
-  font-family: 'Inconsolata', monospace;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: white;
-}
