+++ src/constants/navData.ts
@@ -1,0 +1,22 @@
+export const NAV_DATA = [
+  {
+    to: '/',
+    title: 'Home',
+    id: 'home',
+  },
+  {
+    to: '/search-book',
+    title: 'Search',
+    id: 'search',
+  },
+  {
+    to: '/random-book',
+    title: 'Random Book',
+    id: 'random-book',
+  },
+  {
+    to: '/favorites',
+    title: 'Favorites',
+    id: 'favorites',
+  },
+]
